requirements.txt
pandas
numpy
faker
python-dotenv
azure-storage-blob
mcp